+++ functions/GeneralFunctions/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+ ...require('./commandError'), 
+ ...require('./pad')}
