+++ functions/GeneralFunctions/pad.js
@@ -1,0 +1,10 @@
+const pad = (pad, str, padLeft) => {
+ if (typeof str === "undefined") return pad;
+ if (padLeft) {
+   return (pad + str).slice(-pad.length);
+ } else {
+   return (str + pad).substring(0, pad.length);
+ }
+}
+
+module.exports = {pad}
