+++ ozmacommands/index.js
@@ -1,0 +1,9 @@
+module.exports = {
+ ...require('./elementalrooms'), 
+ ...require('./ozmaplz'),
+ ...require('./plusone'),
+ ...require('./portals'),
+ ...require('./schedule'),
+ ...require('./waymarks'),
+ ...require('./cancel'),
+ ...require('./noplusone')}
